Styles UI to feel like Shivers, adds sounds, cleans up x64 from files, version set from one location.
