+++ Shivers Randomizer_x64/App.xaml
@@ -1,7 +1,7 @@
-﻿<Application x:Class="Shivers_Randomizer_x64.App"
+﻿<Application x:Class="Shivers_Randomizer.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
     <Application.Resources>
-         
+        <FontFamily x:Key="Shivers">./resources/ShiversFont.fon#ShiversFont</FontFamily>
     </Application.Resources>
 </Application>
