--- Shivers Randomizer_x64/AttachPopup_x64.xaml
@@ -1,17 +1,0 @@
-﻿<Window x:Class="Shivers_Randomizer_x64.AttachPopup_x64"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Shivers_Randomizer_x64"
-        mc:Ignorable="d"
-        Title="Attach to Process" Height="400" Width="400" Icon="/resources/Shivers-Icon.ico">
-    <Grid>
-        <Button x:Name="button_GetProcessList" Content="Get Processes" HorizontalAlignment="Left" Margin="50,337,0,0" VerticalAlignment="Top" Width="80" Click="Button_GetProcessList_Click" />
-        <Button x:Name="button_Clear" Content="Clear" HorizontalAlignment="Left" Margin="150,337,0,0" VerticalAlignment="Top" Width="80" Click="Button_Clear_Click"/>
-        <Button x:Name="button_Attach" Content="Attach" HorizontalAlignment="Left" Margin="250,337,0,0" VerticalAlignment="Top" Width="80" Click="Button_Attach_Click"/>
-        <ListBox x:Name="listBox_Process_List" HorizontalAlignment="Left" Height="292" Margin="10,10,0,0" VerticalAlignment="Top" Width="372"/>
-
-        <Label x:Name="label_Feedback" Content="" HorizontalAlignment="Left" Margin="10,307,0,0" VerticalAlignment="Top" Width="372"/>
-    </Grid>
-</Window>
