+++ Shivers Randomizer_x64/Overlay.xaml
@@ -1,0 +1,18 @@
+﻿<Window x:Class="Shivers_Randomizer.Overlay"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:Shivers_Randomizer"
+        mc:Ignorable="d"
+        Title="Overlay"
+        SizeToContent="WidthAndHeight"
+        AllowsTransparency="true"
+        WindowStyle="None"
+        Background="Transparent"
+        Topmost="True"
+        ShowInTaskbar="False">
+    <Grid Margin="10,10,0,0">
+        <Label Name="labelOverlay" Content="Test Overlay Text"  HorizontalAlignment="Center" VerticalAlignment="Top" Foreground="Lime"/>
+    </Grid>
+</Window>
