--- Shivers Randomizer_x64/Overlay_x64.xaml
@@ -1,12 +1,0 @@
-﻿<Window x:Class="Shivers_Randomizer_x64.Overlay_x64"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Shivers_Randomizer_x64"
-        mc:Ignorable="d"
-        Title="Overlay" Height="300" Width="400" AllowsTransparency="true" WindowStyle="None" Background="Transparent" Topmost="True" ShowInTaskbar="False">
-    <Canvas>
-        <Label Name="labelOverlay" Content="Test Overlay Text"  HorizontalAlignment="Center" VerticalAlignment="Top" Foreground="Lime" Canvas.Left="10" Canvas.Top="10"/>
-    </Canvas>
-</Window>
