+++ Shivers Randomizer_x64/Properties/launchSettings.json
@@ -1,0 +1,7 @@
+{
+  "profiles": {
+    "Shivers Randomizer": {
+      "commandName": "Project"
+    }
+  }
+}
