+++ Shivers Randomizer_x64/room_randomizer/Edge.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace Shivers_Randomizer_x64.room_randomizer;
+namespace Shivers_Randomizer.room_randomizer;
 public record Edge
 (
     int First,
