+++ Shivers Randomizer_x64/room_randomizer/Move.cs
@@ -1,4 +1,4 @@
-﻿namespace Shivers_Randomizer_x64.room_randomizer;
+﻿namespace Shivers_Randomizer.room_randomizer;
 
 public record Move
 (
